Remove Board class and moved functions to Game
Having the extra class in the mix didn't seem to be adding any value.

Now that I need to calculate possible collisions in the moves, having
the move functions have access to the available VehiclePlacements seems
like the better approach. Still not 100% sure about this, but will go
with it for now.