Add test for horizontal collision.
In doing so I realiazed that the placement object returned by the get is
invalid after the move as that instance is removed from the list. The
more I think about it, the more I think maybe just the car should be
passed instead of deriving the location in the test.